requirements.txt
ccxt>=4.0.0
requests>=2.31.0
numpy>=1.26.0
pandas>=2.1.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
psutil>=5.9.0
